Added Euler Option
I have added preprocessor directive to define whether to use RK4
method, if set to 0 this uses the euler forward difference method.

Relevant arrays are only allocated depending on option defined. Thus it
should save on RAM and run-time to use the Euler method, but may not
allow as large a timestep.

Tested for dt=0.01 and agrees well up to T=20.